Patch config files to render tests correctly
A trailing comma in `tsconfig.json` was breaking the build in a way that
wasn't immediately apparent what was wrong. In addition, the jest config
file wasn't pointing to the `tsconfig.json` file in `./src/client`
causing `jest` and/or `ts-jest` to go looking trying to find one by
itself. It then found one in he project root (the wrong one), used that,
which then failed to resolve or parse TSX and app files correctly.